Enable several more clang-tidy checks. Fix a set of 'loop variable too small' errors. (#17690)
* enable more clang-tidy checks

* Enable one more check

* More enabled tests

* Enable more checks and make codechanges to make it pass

* Restyle

* Fix compilation

* One more compile failure

* Update more tidy errors in tests

* Mock attribute storage compile fixes

* Add message to static assert

* Undo emberAfClusterCount cast to uint8_t

* More changes for emberAfClusterCount returning uint8_t

* more updates for cluster count

* Updated code for iteration over array size

* one more review update

* one more review update

* Attempt to fix some non-null asserts by tidy

* Adding one more nullable marker

* Adding one more nullable marker

* Fix nullable marker syntax for darwin compile

* Update nullability markers again, this time darwin passes locally for all bridges

* Disable constructor nullability linter for test command constructors as they seem broken

* Move nolint for nullability into darwin-tool rather than main chip-tool

+++ src/setup_payload/Base38Encode.cpp
@@ -55,7 +55,7 @@ CHIP_ERROR base38Encode(ByteSpan in_buf, MutableCharSpan & out_buf)
 
         size_t bytesInChunk = (in_buf_len >= kMaxBytesSingleChunkLen) ? kMaxBytesSingleChunkLen : in_buf_len;
 
-        for (uint8_t byte_idx = 0; byte_idx < bytesInChunk; byte_idx++)
+        for (size_t byte_idx = 0; byte_idx < bytesInChunk; byte_idx++)
         {
             value += static_cast<uint32_t>(in_buf_ptr[byte_idx] << (8 * byte_idx));
         }
